requirements.txt
dotenv
google-adk
litellm
vertexai
psycopg2-binary
asyncpg
greenlet
fastapi
sqlalchemy
uvicorn
python-multipart